CW2_TrailAPI/requirements.txt
asgiref==3.8.1
Django==5.0.9
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
mssql-django==1.5
packaging==24.1
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
wheel